feat: add custom logo
